+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema[7.1].define(version: 2025_12_24_024408) do
+ActiveRecord::Schema[7.1].define(version: 2025_12_24_055522) do
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
 
@@ -125,6 +125,7 @@ ActiveRecord::Schema[7.1].define(version: 2025_12_24_024408) do
     t.datetime "updated_at", null: false
     t.jsonb "reward_payload", default: {}
     t.string "kind"
+    t.index "((reward_payload ->> 'earned_date'::text)), ((reward_payload ->> 'level'::text))", name: "index_rewards_unique_day_level", unique: true, where: "((kind)::text = ANY ((ARRAY['earned'::character varying, 'redeemed'::character varying])::text[]))"
     t.index ["goal_id"], name: "index_rewards_on_goal_id"
   end